product_catalog_app/requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
sqlparse==0.5.3
gunicorn==23.0.0
mysqlclient==2.1.1
python-dotenv==1.0.0
